% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{LDL_CAD}
\alias{LDL_CAD}
\title{Example dataset}
\format{
A data frame with 100 rows and 5 variables:
\describe{
  \item{rsID}{SNP identifier}
  \item{beta.exposure}{Beta effect on exposure}
  \item{beta.outcome}{Beta effect on outcome}
  \item{se.exposure}{Standard error of beta.exposure}
  \item{se.outcome}{Standard error of beta.outcome}
  \item{mr_keep}{whether to include this instruments to estimate causal effect}
}
}
\source{
Available at: https://pmc.ncbi.nlm.nih.gov/articles/PMC6124628/
}
\usage{
LDL_CAD
}
\description{
This summary statistics aims to estimate the causal effect of low-density lipoprotein (LDL) cholesterol on coronary artery disease (CAD) risk, using eight SNPs with strong biological
links to LDL cholesterol.
}
\keyword{datasets}
