% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_F_bar.R
\name{MR_F_bar}
\alias{MR_F_bar}
\title{Calculate mean F statistic}
\usage{
MR_F_bar(beta.exposure, se.exposure)
}
\arguments{
\item{beta.exposure}{A vector of instrument-exposure association estimates.}

\item{se.exposure}{A vector of standard error estimates for the set of instrument-exposure association estimates.}
}
\value{
Returns the F bar statistic value.
}
\description{
Calculate mean F statistic to capture the dilution in the inverse variance weighted estimate for a set of instrument-exposure effects.
}
\examples{
F.bar <- MR_F_bar(beta.exposure = SNP_LDL_assoc, se.exposure = SNP_LDL_assoc_SE)
}
