% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{MR_scatter_estimates}
\alias{MR_scatter_estimates}
\title{Construct a scatter plot of instrument effects and causal effect estimates.}
\usage{
MR_scatter_estimates(
  beta.exposure,
  beta.outcome,
  se.exposure,
  se.outcome,
  b,
  beta.intercepts,
  title_text,
  xlab,
  ylab,
  methods,
  xsize = 22,
  ysize = 22,
  titlesize = 24,
  opacity = 1
)
}
\arguments{
\item{beta.exposure}{A vector of instrument-exposure association estimates.}

\item{beta.outcome}{A vector of instrument-outcome association estimates from a common instrument set.}

\item{se.exposure}{A vector of standard error estimates for the set of instrument-exposure association estimates.}

\item{se.outcome}{A vector of stand error estimates for the set of instrument-outcome association estimates.}

\item{b}{A vector of slope estimates for a given exposure-outcome pair.}

\item{beta.intercepts}{A vector of intercepts corresponding to the set of causal effect estimates given. Only the MR-Egger intercept should be non-zero.}

\item{title_text}{A string object specifying the title of the customised coverage plot.}

\item{xlab}{A string object specifying the x-axis title for the scatter plot.}

\item{ylab}{A string object specifying the y-axis title for the scatter plot.}

\item{methods}{A vector detailing the name of chosen causal effect estimators.}

\item{xsize}{A numeric object specifying the desired size of the x-axis tick markers. The default value is 22.}

\item{ysize}{A numeric object specifying the desired size of the y-axis tick markers. The default value is 22.}

\item{titlesize}{A numeric object specifying the desired size of the plot title. The default value is 24.}

\item{opacity}{A numeric object specifying the opacity of the scatter plot markers (between 0 and 1). The default value is 1.}
}
\value{
Returns a plotly plot object.
}
\description{
Plot instrument-exposure effects and instrument-outcome effects (with corresponding 95% confidence intervals). Overlay causal effect estimates as slopes to allow visual inspection of similarity between causal effect estimates.
}
\examples{
SNP.plot <- MR_2s_scatter_auto(beta.exposure = SNP_LDL.beta, beta.outcome = SNP_CHD.beta, se.exposure = SNP_LDL.beta.se, se.outcome = SNP_CHD.beta.se, b = ests.beta, beta.intercepts = c(0, -0.01, 0, 0), title.text = "Causal effect estimates of LDL cholesterol on coronary heart disease risk", xlab = "SNP-LDL effect", ylab = "SNP-CHD effect", methods = c("IVW", "MRE", "WM", "MBE"), xsize = 25, ysize = 25, titlesize = 25)
}
